Language.PlutusTx.Coordination.Contracts.CrowdFunding

Description

Crowdfunding contract implemented using the [[Plutus]] interface. This is the fully parallel version that collects all contributions in a single transaction. This is, of course, limited by the maximum number of inputs a transaction can have.

Campaign parameters

data Campaign

A crowdfunding campaign.

Constructors

Campaign

Fields

campaignDeadline :: Slot

campaignTarget :: Value

campaignCollectionDeadline :: Slot

campaignOwner :: CampaignActor

type CampaignActor = PubKey

Functionality for campaign contributors

contribute :: (WalletAPI m, WalletDiagnostics m) => Campaign -> Value -> m ()

Contribute funds to the campaign (contributor)

Functionality for campaign owners

collect :: (WalletAPI m, WalletDiagnostics m) => Campaign -> m ()

Register a [[EventHandler]] to collect all the funds of a campaign

campaignAddress :: Campaign -> Address

The address of a [[Campaign]]