Language.PlutusTx.Coordination.Contracts.Future

Description

A futures contract in Plutus. This example illustrates three concepts. 1. Maintaining a margin (a kind of deposit) during the duration of the contract to protect against breach of contract (see note [Futures in Plutus]) 2. Using oracle values to obtain current pricing information (see note [Oracles] in Language.PlutusTx.Coordination.Contracts) 3. Using the redeemer script to model actions that the participants in the contract may take.

Data types

data Future

Basic data of a futures contract. Future contains all values that do not change during the lifetime of the contract.

Constructors

Future

Fields

futureDeliveryDate :: Slot

futureUnits :: Int

futureUnitPrice :: Value

futureInitialMargin :: Value

futurePriceOracle :: PubKey

futureMarginPenalty :: Value

How much a participant loses if they fail to make the required margin payments.

data FutureData

The current "state" of the futures contract. FutureData contains values that may change during the lifetime of the contract. This is the data script.

Constructors

FutureData

Fields

futureDataLong :: PubKey

Holder of the long position (buyer)

futureDataShort :: PubKey

Holder of the short position (seller)

futureDataMarginLong :: Value

Current balance of the margin account of the long position

futureDataMarginShort :: Value

Current balance of the margin account of the short position

data FutureRedeemer

Actions that either participant may take. This is the redeemer script.

Constructors

AdjustMargin
Make a margin payment

Settle (OracleValue Value)
Settle the contract

Actions

initialise

Arguments

:: (MonadError WalletAPIError m, WalletAPI m)

=> PubKey
Identity of the holder of the long position

-> PubKey
Identity of the holder of the short position

-> Future

-> m ()

Initialise the futures contract by paying the initial margin.

settle :: (MonadError WalletAPIError m, WalletAPI m) => [TxOutRef] -> Future -> FutureData -> OracleValue Value -> m ()

Close the position by extracting the payment

settleEarly :: (MonadError WalletAPIError m, WalletAPI m) => [TxOutRef] -> Future -> FutureData -> OracleValue Value -> m ()

Settle the position early if a margin payment has been missed.

adjustMargin :: (MonadError WalletAPIError m, WalletAPI m) => [TxOutRef] -> Future -> FutureData -> Value -> m ()

Script

validatorScript :: Future -> ValidatorScript