Language.PlutusTx.Coordination.Contracts.Swap

Documentation

data Swap

A swap is an agreement to exchange cashflows at future dates. To keep things simple, this is an interest rate swap (meaning that the cashflows are interest payments on the same principal amount but with two different interest rates, of which one is fixed and one is floating (varying with time)) with only a single payment date.

At the beginning of the contract, the fixed rate is set to the expected future value of the floating rate (so if the floating rate behaves as expected, the two payments will be exactly equal).

Constructors

Swap

Fields

swapNotionalAmt :: !Value

swapObservationTime :: !Slot

swapFixedRate :: !(Ratio Int)

Interest rate fixed at the beginning of the contract

swapFloatingRate :: !(Ratio Int)

Interest rate whose value will be observed (by an oracle) on the day of the payment

swapMargin :: !Value

Margin deposited at the beginning of the contract to protect against default (one party failing to pay)

swapOracle :: !PubKey

Public key of the oracle (see note [Oracles] in [[Language.PlutusTx.Coordination.Contracts]])

swapValidator :: Swap -> ValidatorScript

Validator script for the two transactions that initialise the swap. See note [Swap Transactions] See note [Contracts and Validator Scripts] in Language.Plutus.Coordination.Contracts