Language.PlutusTx.Coordination.Contracts.Vesting

Description

Vesting scheme as a PLC contract

Documentation

data Vesting

A vesting scheme consisting of two tranches. Each tranche defines a date (slot) after which an additional amount of money can be spent.

Constructors

Vesting

Fields

vestingTranche1 :: VestingTranche

vestingTranche2 :: VestingTranche

vestingOwner :: PubKey

data VestingTranche

Tranche of a vesting scheme.

Constructors

VestingTranche

Fields

vestingTrancheDate :: Slot

vestingTrancheAmount :: Value

data VestingData

Data script for vesting utxo

Constructors

VestingData

Fields

vestingDataHash :: ValidatorHash

Hash of the validator script

vestingDataPaidOut :: Value

How much of the vested value has already been retrieved

vestFunds :: (MonadError WalletAPIError m, WalletAPI m) => Vesting -> Value -> m VestingData

Lock some funds with the vesting validator script and return a [[VestingData]] representing the current state of the process

retrieveFunds

Arguments

:: (Monad m, WalletAPI m)

=> Vesting

-> VestingData
Value that has already been taken out

-> TxOutRef
Transaction output locked by the vesting validator script

-> Value
Value we want to take out now

-> m VestingData

Retrieve some of the vested funds.

validatorScript :: Vesting -> ValidatorScript

totalAmount :: Vesting -> Value

The total amount of money vested

validatorScriptHash :: Vesting -> ValidatorHash