Plutus.Contracts.Crowdfunding

Description

Crowdfunding contract implemented using the [[Plutus]] interface. This is the fully parallel version that collects all contributions in a single transaction. This is, of course, limited by the maximum number of inputs a transaction can have.

Campaign parameters

data Campaign

A crowdfunding campaign.

Constructors

Campaign

Fields

campaignDeadline :: POSIXTime

The date by which the campaign funds can be contributed.

campaignCollectionDeadline :: POSIXTime

The date by which the campaign owner has to collect the funds

campaignOwner :: PaymentPubKeyHash

Public key of the campaign owner. This key is entitled to retrieve the funds if the campaign is successful.

type CrowdfundingSchema = Endpoint "schedule collection" () .\/ Endpoint "contribute" Contribution

crowdfunding :: Campaign -> Contract () CrowdfundingSchema ContractError ()

The crowdfunding contract for the Campaign.

theCampaign :: POSIXTime -> Campaign

A sample campaign

Functionality for campaign contributors

contribute :: Campaign -> Promise () CrowdfundingSchema ContractError ()

The "contribute" branch of the contract for a specific Campaign. Exposes an endpoint that allows the user to enter their public key and the contribution. Then waits until the campaign is over, and collects the refund if the funding was not collected.

newtype Contribution

Constructors

Contribution

Fields

contribValue :: Value

how much to contribute

Functionality for campaign owners

scheduleCollection :: Campaign -> Promise () CrowdfundingSchema ContractError ()

The campaign owner's branch of the contract for a given Campaign. It watches the campaign address for contributions and collects them if the funding goal was reached in time.

campaignAddress :: Campaign -> ValidatorHash

The address of a [[Campaign]]

Validator script

contributionScript :: Campaign -> Validator

The validator script that determines whether the campaign owner can retrieve the funds or the contributors can claim a refund.

mkValidator :: Campaign -> PaymentPubKeyHash -> CampaignAction -> ScriptContext -> Bool

The validator script is of type CrowdfundingValidator, and is additionally parameterized by a Campaign definition. This argument is provided by the Plutus client, using applyCode. As a result, the Campaign definition is part of the script address, and different campaigns have different addresses. The Campaign{..} syntax means that all fields of the Campaign value are in scope (for example campaignDeadline in l. 70).

mkCampaign :: POSIXTime -> POSIXTime -> Wallet -> Campaign

Construct a Campaign value from the campaign parameters, using the wallet's public key.

data CampaignAction

Action that can be taken by the participants in this contract. A value of CampaignAction is provided as the redeemer. The validator script then checks if the conditions for performing this action are met.

Constructors

Collect

Refund

collectionRange :: Campaign -> ValidityInterval POSIXTime

The 'ValidityInterval POSIXTime' during which the funds can be collected

refundRange :: Campaign -> ValidityInterval POSIXTime

The 'ValidityInterval POSIXTime' during which a refund may be claimed

Traces

startCampaign :: EmulatorTrace (ContractHandle () CrowdfundingSchema ContractError)

Call the "schedule collection" endpoint and instruct the campaign owner's wallet (wallet 1) to start watching the campaign address.

makeContribution :: Wallet -> Value -> EmulatorTrace ()

Call the "contribute" endpoint, contributing the amount from the wallet

successfulCampaign :: EmulatorTrace ()

Run a successful campaign with contributions from wallets 2, 3 and 4.