Plutus.Contracts.Escrow

Description

A general-purpose escrow contract in Plutus

Documentation

The escrow contract implements the exchange of value between multiple parties. It is defined by a list of targets (public keys and script addresses, each associated with a value). It works similar to the crowdfunding contract in that the contributions can be made independently, and the funds can be unlocked only by a transaction that pays the correct amount to each target. A refund is possible if the outputs locked by the contract have not been spent by the deadline. (Compared to the crowdfunding contract, the refund policy is simpler because here because there is no "collection period" during which the outputs may be spent after the deadline has passed. This is because we're assuming that the participants in the escrow contract will make their deposits as quickly as possible after agreeing on a deal)

The contract supports two modes of operation, manual and automatic. In manual mode, all actions are driven by endpoints that exposed via payEp redeemEp and refundEp. In automatic mode, the pay, redeem and refundactions start immediately. This mode is useful when the escrow is called from within another contract, for example during setup (collection of the initial deposits).

data Escrow

data EscrowError

Constructors

RedeemFailed RedeemFailReason

RefundFailed

EContractError ContractError

class AsEscrowError r where

Minimal complete definition

_EscrowError

Methods

_EscrowError :: Prism' r EscrowError

_RedeemFailed :: Prism' r RedeemFailReason

_RefundFailed :: Prism' r ()

_EContractError :: Prism' r ContractError

data EscrowParams d

Definition of an escrow contract, consisting of a deadline and a list of targets

Constructors

EscrowParams

Fields

escrowDeadline :: POSIXTime

Latest point at which the outputs may be spent.

escrowTargets :: [EscrowTarget d]

Where the money should go. For each target, the contract checks that the output mkTxOutput of the target is present in the spending transaction.

data EscrowTarget d

Defines where the money should go. Usually we have `d = Datum` (when defining EscrowTarget values in off-chain code). Sometimes we have `d = DatumHash` (when checking the hashes in on-chain code)

Constructors

PaymentPubKeyTarget PaymentPubKeyHash Value

ScriptTarget ValidatorHash d Value

payToScriptTarget :: ValidatorHash -> Datum -> Value -> EscrowTarget Datum

An EscrowTarget that pays the value to a script address, with the given data script.

payToPaymentPubKeyTarget :: PaymentPubKeyHash -> Value -> EscrowTarget d

An EscrowTarget that pays the value to a public key address.

targetTotal :: EscrowParams d -> Value

The total Value that must be paid into the escrow contract before it can be unlocked

escrowContract :: EscrowParams Datum -> Contract () EscrowSchema EscrowError ()

payRedeemRefund :: forall w s. EscrowParams Datum -> Value -> Contract w s EscrowError (Either RefundSuccess RedeemSuccess)

Pay some money into the escrow contract. Then release all funds to their specified targets if enough funds were deposited before the deadline, or reclaim the contribution if the goal has not been met.

typedValidator :: EscrowParams Datum -> TypedValidator Escrow

Actions

pay

Arguments

:: forall w s e. AsContractError e

=> TypedValidator Escrow
The instance

-> EscrowParams Datum
The escrow contract

-> Value
How much money to pay in

-> Contract w s e TxId

Pay some money into the escrow contract.

payEp :: forall w s e. (HasEndpoint "pay-escrow" Value s, AsEscrowError e) => EscrowParams Datum -> Promise w s e TxId

pay with an endpoint that gets the owner's public key and the contribution.

redeem :: forall w s e. AsEscrowError e => TypedValidator Escrow -> EscrowParams Datum -> Contract w s e RedeemSuccess

Redeem all outputs at the contract address using a transaction that has all the outputs defined in the contract's list of targets.

redeemEp :: forall w s e. (HasEndpoint "redeem-escrow" () s, AsEscrowError e) => EscrowParams Datum -> Promise w s e RedeemSuccess

redeem with an endpoint.

refund :: forall w s. TypedValidator Escrow -> EscrowParams Datum -> Contract w s EscrowError RefundSuccess

Claim a refund of the contribution.

refundEp :: forall w s. HasEndpoint "refund-escrow" () s => EscrowParams Datum -> Promise w s EscrowError RefundSuccess

refund with an endpoint.

data RedeemFailReason

Constructors

DeadlinePassed

NotEnoughFundsAtAddress

newtype RedeemSuccess

Constructors

RedeemSuccess TxId

newtype RefundSuccess

Constructors

RefundSuccess TxId

type EscrowSchema = (Endpoint "pay-escrow" Value .\/ Endpoint "redeem-escrow" ()) .\/ Endpoint "refund-escrow" ()

Exposed for test endpoints

data Action

Constructors

Redeem

Refund

Coverage

covIdx :: CoverageIndex