Plutus.Contracts.Game.Alonzo

Description

A guessing game. A simplified version of GameStateMachine not using StateMachine and using yieldUnbalancedTx for balancing, signing and submitting transactions.

Currently, remote wallets (anything other than WBE) can only handles yieldUnbalancedTx requests, and not balanceTx, signTx and submitTx requests.

This is the "Alonzo version" of the game, it uses features up to the Alonzo era.

Documentation

contract :: AsContractError e => Contract () GameSchema e ()

data GameParam

Datatype for creating a parameterized validator.

Constructors

GameParam

Fields

gameParamPayeePkh :: PaymentPubKeyHash

Payment public key hash of the wallet locking some funds

gameParamStartTime :: POSIXTime

Starting time of the game

type GameSchema = Endpoint "lock" LockArgs .\/ Endpoint "guess" GuessArgs

data LockArgs

Arguments for the "lock" endpoint

Constructors

LockArgs

Fields

lockArgsGameParam :: GameParam

The parameters for parameterizing the validator.

lockArgsSecret :: String

The secret

lockArgsValue :: Value

Value that is locked by the contract initially

data GuessArgs

Arguments for the "guess" endpoint

Constructors

GuessArgs

Fields

guessArgsGameParam :: GameParam

The parameters for parameterizing the validator.

guessArgsSecret :: String

The guess

Scripts

gameInstance :: GameParam -> TypedValidator Game

mkValidator :: GameParam -> HashedString -> ClearString -> ScriptContext -> Bool

The validation function (Datum -> Redeemer -> ScriptContext -> Bool)

The GameParam parameter is not used in the validation. It is meant to parameterize the script address depending based on the value of GaramParam.

Address

gameAddress :: GameParam -> CardanoAddress

The address of the game (the hash of its validator script)

covIdx :: GameParam -> CoverageIndex