Plutus.Contracts.GameStateMachine

Description

A guessing game that

Uses a state machine to keep track of the current secret word

Uses a token to keep track of who is allowed to make a guess

Documentation

contract :: Contract () GameStateMachineSchema GameError ()

Top-level contract, exposing both endpoints.

typedValidator :: GameParam -> TypedValidator GameStateMachine

data GameParam

Datatype for creating a parameterized validator.

Constructors

GameParam

Fields

gameParamPayeePkh :: Address

Payment address of the wallet locking some funds

gameParamStartTime :: POSIXTime

Starting time of the game

data GameState

State of the guessing game

Constructors

Initialised MintingPolicyHash TokenName HashedString
Initial state. In this state only the MintTokens action is allowed.

Locked MintingPolicyHash TokenName HashedString
Funds have been locked. In this state only the Guess action is allowed.

Finished
All funds were unlocked.

data GameInput

Inputs (actions)

Constructors

MintToken
Mint the "guess" token

Guess Address ClearString HashedString Value
Make a guess, extract the funds, and lock the remaining funds using a new secret word.

data GuessToken

The token that represents the right to make a guess

mkValidator :: GameParam -> ValidatorType GameStateMachine

mintingPolicy :: GameParam -> MintingPolicy

data LockArgs

Arguments for the "lock" endpoint

Constructors

LockArgs

Fields

lockArgsGameParam :: GameParam

The parameters for parameterizing the validator.

lockArgsSecret :: SecretArgument String

The secret

lockArgsValue :: Value

Value that is locked by the contract initially

data GuessArgs

Arguments for the "guess" endpoint

Constructors

GuessArgs

Fields

guessArgsGameParam :: GameParam

The parameters for parameterizing the validator.

guessTokenTarget :: Address

The recipient of the guess token

guessArgsOldSecret :: String

The guess

guessArgsNewSecret :: SecretArgument String

The new secret

guessArgsValueTakenOut :: Value

How much to extract from the contract

type GameStateMachineSchema = Endpoint "lock" LockArgs .\/ Endpoint "guess" GuessArgs

The schema of the contract. It consists of the two endpoints "lock" and "guess" with their respective argument types.

data GameError

token :: MintingPolicyHash -> TokenName -> Value

covIdx :: CoverageIndex