Plutus.Contracts.MultiSigStateMachine

Description

A multisig contract written as a state machine.

Documentation

The n-out-of-m multisig contract works like a joint account of m people, requiring the consent of n people for any payments. In the smart contract the signatories are represented by public keys, and approval takes the form of signatures on transactions.

The multisig contract in MultiSig expects n signatures on a single transaction. This requires an off-chain communication channel. The multisig contract implemented in this module uses a proposal system that allows participants to propose payments and attach their signatures to proposals over a period of time, using separate transactions. All contract state is kept on the chain so there is no need for off-chain communication.

data Params

Constructors

Params

Fields

mspSignatories :: [PaymentPubKeyHash]

Public keys that are allowed to authorise payments

mspRequiredSigs :: Integer

How many signatures are required for a payment

data Payment

A proposal for making a payment under the multisig scheme.

Constructors

Payment

Fields

paymentAmount :: Value

How much to pay out

paymentRecipient :: Address

Address to pay the value to

paymentDeadline :: POSIXTime

Time until the required amount of signatures has to be collected.

data State s

mkValidator :: Params -> ValidatorType MultiSigSym

typedValidator :: Params -> TypedValidator MultiSigSym

data MultiSigError

Constructors

MSContractError ContractError

MSStateMachineError SMContractError

type MultiSigSchema = (((Endpoint "propose-payment" Payment .\/ Endpoint "add-signature" ()) .\/ Endpoint "cancel-payment" ()) .\/ Endpoint "pay" ()) .\/ Endpoint "lock" Value

contract :: (AsContractError e, AsSMContractError e) => Params -> Contract () MultiSigSchema e ()