Plutus.Contracts.Prism.Unlock

Description

Two sample that unlock some funds by presenting the credentials. * subscribeSTO uses the credential to participate in an STO * unlockExchange uses the credential to take ownership of funds that were locked by an exchange.

STO

data STOSubscriber

Constructors

STOSubscriber

Fields

wCredential :: Credential

wSTOIssuer :: PaymentPubKeyHash

wSTOTokenName :: TokenName

wSTOAmount :: Integer

type STOSubscriberSchema = Endpoint "sto" STOSubscriber

subscribeSTO :: forall w s. HasEndpoint "sto" STOSubscriber s => Contract w s UnlockError ()

Obtain a token from the credential manager app, then participate in the STO

Exchange

type UnlockExchangeSchema = Endpoint "unlock from exchange" Credential

unlockExchange :: forall w s. HasEndpoint "unlock from exchange" Credential s => Contract w s UnlockError ()

Obtain a token from the credential manager app, then use it to unlock funds that were locked by an exchange.

Errors etc.

data UnlockError

Constructors

WithdrawEndpointError ContractError

WithdrawTxError ContractError

WithdrawPkError ContractError

GetCredentialStateMachineError SMContractError

GetCredentialTransitionError (InvalidTransition IDState IDAction)

UnlockExchangeTokenAccError TokenAccountError

UnlockMkTxError MkTxError