Plutus.Contracts.Uniswap.Pool

Documentation

calculateAdditionalLiquidity :: Amount A -> Amount B -> Amount Liquidity -> Amount A -> Amount B -> Amount Liquidity

calculateInitialLiquidity :: Amount A -> Amount B -> Amount Liquidity

The initial liquidity is 'ceil( sqrt(x*y) )' where x is the amount of 'Coin A' and y the amount of 'Coin B'. See Eq. 13 of the Uniswap v2 paper.

calculateRemoval :: Amount A -> Amount B -> Amount Liquidity -> Amount Liquidity -> (Amount A, Amount B)

See Definition 3 of https://github.com/runtimeverification/verified-smart-contracts/blob/c40c98d6ae35148b76742aaaa29e6eaa405b2f93/uniswap/x-y-k.pdf.

checkSwap :: Amount A -> Amount B -> Amount A -> Amount B -> Bool

A swap is valid if the fee is computed correctly, and we're swapping some positive amount of A for B. See: https://uniswap.org/whitepaper.pdf Eq (11) (Page 7.)

lpTicker :: LiquidityPool -> TokenName

Generate a unique token name for this particular pool; based on the tokens it exchanges. This should be such that looking for a pool exchanging any two tokens always yields a unique name.