Plutus.Contracts.Uniswap.Trace

Description

Example trace for the uniswap contract

Documentation

uniswapTrace :: EmulatorTrace ()

Set up a liquidity pool and call the "add" endpoint

setupTokens :: Contract (Maybe (Last OneShotCurrency)) CurrencySchema CurrencyError ()

Create some sample tokens and distribute them to the emulated wallets

tokenNames :: [TokenName]

wallets :: [Wallet]

increaseTransactionLimits :: Params -> Params

increaseTransactionLimitsOpts :: CheckOptions -> CheckOptions