Plutus.Contracts.Auction

Documentation

data AuctionState

The states of the auction

Constructors

Ongoing HighestBid

Finished HighestBid

data AuctionInput

Transition between auction states

Constructors

Bid

Fields

newBid :: Ada

newBidder :: Address

Payout

type BuyerSchema = Endpoint "bid" Ada

type SellerSchema = EmptySchema

data AuctionParams

Definition of an auction

Constructors

AuctionParams

Fields

apOwner :: Address

Current owner of the asset. This is where the proceeds of the auction will be sent.

apAsset :: Value

The asset itself. This value is going to be locked by the auction script output.

apEndTime :: POSIXTime

When the time window for bidding ends.

data HighestBid

Constructors

HighestBid

Fields

highestBid :: Ada

highestBidder :: Address

auctionBuyer :: ThreadToken -> AuctionParams -> Contract AuctionOutput BuyerSchema AuctionError ()

auctionSeller :: Value -> POSIXTime -> Contract AuctionOutput SellerSchema AuctionError ()

Client code for the seller

data AuctionOutput

Observable state of the auction app

Constructors

AuctionOutput

Fields

auctionState :: Last AuctionState

auctionThreadToken :: Last ThreadToken

data AuctionError

Constructors

StateMachineContractError SMContractError
State machine operation failed

AuctionContractError ContractError
Endpoint, coin selection, etc. failed

data ThreadToken

getThreadToken :: forall e w (schema :: Row Type). AsSMContractError e => Contract w schema e ThreadToken

Create a thread token. The thread token contains a reference to an unspent output of the wallet, so it needs to used with mkStateMachine immediately, and the machine must be initialised, to prevent the output from getting spent in the mean time.

covIdx :: CoverageIndex