Plutus.Contracts.Currency

Description

Implements a custom currency with a minting policy that allows the minting of a fixed amount of units.

Documentation

data OneShotCurrency

A currency that can be created exactly once

Constructors

OneShotCurrency

Fields

curRefTransactionOutput :: (TxId, Integer)

Transaction input that must be spent when the currency is minted.

curAmounts :: Map TokenName Integer

How many units of each TokenName are to be minted.

type CurrencySchema = Endpoint "Create native token" SimpleMPS

newtype CurrencyError

Constructors

CurContractError ContractError

class AsCurrencyError r where

Minimal complete definition

_CurrencyError

Methods

_CurrencyError :: Prism' r CurrencyError

_CurContractError :: Prism' r ContractError

curPolicy :: OneShotCurrency -> MintingPolicy

Actions etc

mintContract :: forall w s e. AsCurrencyError e => CardanoAddress -> [(TokenName, Integer)] -> Contract w s e OneShotCurrency

mint [(n1, c1), ..., (n_k, c_k)] creates a new currency with k token names, minting c_i units of each token n_i. If k == 0 then no value is minted. A one-shot minting policy script is used to ensure that no more units of the currency can be minted afterwards.

mintedValue :: OneShotCurrency -> Value

The Value minted by the OneShotCurrency contract

currencySymbol :: OneShotCurrency -> CurrencySymbol

Simple minting policy currency

data SimpleMPS

Minting policy for a currency that has a fixed amount of tokens issued in one transaction

Constructors

SimpleMPS

Fields

tokenName :: TokenName

amount :: Integer

mintCurrency :: Promise (Maybe (Last OneShotCurrency)) CurrencySchema CurrencyError OneShotCurrency

Use mintContract to create the currency specified by a SimpleMPS