Plutus.Contracts.Future

Documentation

A futures contract in Plutus. This example illustrates a number of concepts.

Maintaining a margin (a kind of deposit) during the duration of the contract to protect against breach of contract (see note [Futures in Plutus])

Using oracle values to obtain current pricing information (see note [Oracles] in Plutus.Contracts)

Writing contracts as state machines

Using tokens to represent claims on future cash flows

data Future

Basic data of a futures contract. Future contains all values that do not change during the lifetime of the contract.

Constructors

Future

Fields

ftDeliveryDate :: POSIXTime

ftUnits :: Integer

ftUnitPrice :: Value

ftInitialMargin :: Value

ftPriceOracle :: PaymentPubKey

ftMarginPenalty :: Value

How much a participant loses if they fail to make the required margin payments.

data FutureAccounts

The token accounts that represent ownership of the two sides of the future. When the contract is done, payments will be made to these accounts.

Constructors

FutureAccounts

Fields

ftoLong :: Account

The owner of the "long" account (represented by a token)

ftoLongAccount :: ValidatorHash

Address of the TokenAccount validator script for ftoLong. This hash can be derived from ftoLong, but only in off-chain code. We store it here so that we can lift it into on-chain code.

ftoShort :: Account

The owner of the "short" account (represented by a token).

ftoShortAccount :: ValidatorHash

Address of the TokenAccount validator script for ftoShort. The comment on ftoLongAccount applies to this as well.

mkAccounts :: Account -> Account -> FutureAccounts

data FutureError

Constructors

TokenSetupFailed CurrencyError
Something went wrong during the setup of the two tokens

StateMachineError SMContractError

OtherFutureError ContractError

EscrowFailed EscrowError
The escrow that initialises the future contract failed

EscrowRefunded RefundSuccess
The other party didn't make their payment in time so the contract never started.

type FutureSchema = (((Endpoint "initialise-future" (FutureSetup, Role) .\/ Endpoint "join-future" (FutureAccounts, FutureSetup)) .\/ Endpoint "increase-margin" (Value, Role)) .\/ Endpoint "settle-early" (SignedMessage (Observation Value))) .\/ Endpoint "settle-future" (SignedMessage (Observation Value))

data FutureSetup

The data needed to initialise the futures contract.

Constructors

FutureSetup

Fields

shortPK :: PaymentPubKeyHash

Initial owner of the short token

longPK :: PaymentPubKeyHash

Initial owner of the long token

contractStart :: POSIXTime

Start of the futures contract itself. By this time the setup code has to be finished, otherwise the contract is void.

data Role

The two roles involved in the contract.

Constructors

Long

Short

futureContract :: Future -> Contract () FutureSchema FutureError ()

futureStateMachine :: Future -> FutureAccounts -> StateMachine FutureState FutureAction

validator :: Future -> FutureAccounts -> Validator

initialiseFuture :: (HasEndpoint "initialise-future" (FutureSetup, Role) s, AsFutureError e) => Future -> Promise w s e (StateMachineClient FutureState FutureAction)

Initialise the contract by * Generating the tokens for long and short * Setting up an escrow contract for the initial margins * Paying the initial margin for the given role

initialMargin :: Future -> Value

futureAddress :: Future -> FutureAccounts -> Address

tokenFor :: Role -> FutureAccounts -> Value

initialState :: Future -> FutureState

The initial state of the Future contract

typedValidator :: Future -> FutureAccounts -> TypedValidator (StateMachine FutureState FutureAction)

setupTokens :: forall w s e. AsFutureError e => Contract w s e FutureAccounts

Create two unique tokens that can be used for the short and long positions and return a FutureAccounts value for them.

Note that after setupTokens is complete, both tokens will be locked by a public key output belonging to the wallet that ran setupTokens.

setupTokensTrace :: EmulatorTrace ()