Plutus.Contracts.Governance

Description

A basic governance contract in Plutus.

Documentation

When the contract starts it produces a number of tokens that represent voting rights.

Holders of those tokens can propose changes to the state of the contract and vote on them.

After a certain period of time the voting ends and the proposal is rejected or accepted.

contract :: AsGovError e => Params -> Contract () Schema e ()

The main contract for creating a new law and for voting on proposals.

proposalContract :: AsGovError e => Params -> Address -> Proposal -> Contract () EmptySchema e ()

The contract for proposing changes to a law.

data Params

The governace contract parameters.

Constructors

Params

Fields

baseTokenName :: TokenName

The token names that allow voting are generated by adding an increasing number to the base token name. See mkTokenName.

initialHolders :: [Address]

The public key hashes of the initial holders of the voting tokens.

requiredVotes :: Integer

The number of votes in favor required for a proposal to be accepted.

data Proposal

The parameters for the proposal contract.

Constructors

Proposal

Fields

newLaw :: Law

The new contents of the law

tokenName :: TokenName

The name of the voting tokens. Only voting token owners are allowed to propose changes.

votingDeadline :: POSIXTime

The time when voting ends and the votes are tallied.

type Schema = Endpoint "new-law" Law .\/ Endpoint "add-vote" (Address, TokenName, Bool)

The endpoints of governance contracts are

new-law to create a new law and distribute voting tokens

add-vote to vote on a proposal with the name of the voting token and a boolean to vote in favor or against.

mkTokenName :: TokenName -> Integer -> TokenName

Generate a voting token name by tagging on a number after the base token name.

typedValidator :: Params -> TypedValidator GovernanceMachine

mkValidator :: Params -> ValidatorType GovernanceMachine

data GovState

Constructors

GovState

Fields

law :: Law

mph :: MintingPolicyHash

voting :: Maybe Voting

newtype Law

Constructors

Law

Fields

unLaw :: BuiltinByteString

data Voting

Constructors

Voting

Fields

proposal :: Proposal

votes :: Map TokenName Bool

data GovError