Plutus.Contracts.MultiSig

Description

Implements an n-out-of-m multisig contract.

Documentation

data MultiSig

Constructors

MultiSig

Fields

signatories :: [PaymentPubKeyHash]

List of public keys of people who may sign the transaction

minNumSignatures :: Integer

Minimum number of signatures required to unlock the output (should not exceed length signatories)

type MultiSigSchema = Endpoint "lock" (MultiSig, Value) .\/ Endpoint "unlock" (MultiSig, [PaymentPubKeyHash])

contract :: AsContractError e => Contract () MultiSigSchema e ()

lock :: AsContractError e => Promise () MultiSigSchema e ()

Lock some funds in a MultiSig contract.

typedValidator :: MultiSig -> TypedValidator MultiSig

validate :: MultiSig -> () -> () -> ScriptContext -> Bool