Plutus.Contracts.Prism

Unlock (STO)

data STOSubscriber

Constructors

STOSubscriber

Fields

wCredential :: Credential

wSTOIssuer :: PaymentPubKeyHash

wSTOTokenName :: TokenName

wSTOAmount :: Integer

type STOSubscriberSchema = Endpoint "sto" STOSubscriber

data UnlockError

Constructors

WithdrawEndpointError ContractError

WithdrawTxError ContractError

WithdrawPkError ContractError

GetCredentialStateMachineError SMContractError

GetCredentialTransitionError (InvalidTransition IDState IDAction)

UnlockExchangeTokenAccError TokenAccountError

UnlockMkTxError MkTxError

subscribeSTO :: forall w s. HasEndpoint "sto" STOSubscriber s => Contract w s UnlockError ()

Obtain a token from the credential manager app, then participate in the STO

Unlock (exchange)

type UnlockExchangeSchema = Endpoint "unlock from exchange" Credential

unlockExchange :: forall w s. HasEndpoint "unlock from exchange" Credential s => Contract w s UnlockError ()

Obtain a token from the credential manager app, then use it to unlock funds that were locked by an exchange.

Mirror app

type MirrorSchema = Endpoint "issue" CredentialOwnerReference .\/ Endpoint "revoke" CredentialOwnerReference

data CredentialOwnerReference

Reference to a credential tied to a specific owner (public key address). From this, and the public key of the Mirror instance, we can compute the address of the state machine script that locks the token for the owner.

Constructors

CredentialOwnerReference

Fields

coTokenName :: TokenName

coOwner :: Wallet

data MirrorError

Constructors

StateNotFound TokenName PaymentPubKeyHash

SetupError ContractError

MirrorEndpointError ContractError

CreateTokenTxError ContractError

StateMachineError SMContractError

mirror :: (HasEndpoint "revoke" CredentialOwnerReference s, HasEndpoint "issue" CredentialOwnerReference s) => Contract w s MirrorError ()

Credential manager app

data Credential

Named credential issued by a credential authority

Constructors

Credential

Fields

credAuthority :: CredentialAuthority

credName :: TokenName

data UserCredential

A Credential issued to a user (public key address)

Constructors

UserCredential

Fields

ucAddress :: PaymentPubKeyHash

Address of the credential holder

ucCredential :: Credential

The credential

ucToken :: Value

The Value containing a token of the credential (this needs to be included here because token is not available in on-chain code)

newtype CredentialAuthority

Entity that is authorised to mint credential tokens

Constructors

CredentialAuthority

Fields

unCredentialAuthority :: PaymentPubKeyHash

all-in-one

data Role

The roles that we pass to contract.

Constructors

Mirror
The mirror contract

UnlockSTO
The subscribeSTO contract

UnlockExchange
The unlockExchange contract

type PrismSchema = ((MirrorSchema .\/ STOSubscriberSchema) .\/ UnlockExchangeSchema) .\/ Endpoint "role" Role

data PrismError

Constructors

UnlockSTOErr UnlockError

MirrorErr MirrorError

EPError ContractError

UnlockExchangeErr UnlockError

contract :: Contract () PrismSchema PrismError ()

A wrapper around the four prism contracts. This is just a workaround for the emulator, where we can only ever run a single Contract. In the PAB we could simply start all four contracts (credentialManager, mirror, subscribeSTO, subscribeExchange) separately.