Plutus.Contracts.Prism.Mirror

Description

The Atala Mirror application that initialises the state machine

Documentation

type MirrorSchema = Endpoint "issue" CredentialOwnerReference .\/ Endpoint "revoke" CredentialOwnerReference

data CredentialOwnerReference

Reference to a credential tied to a specific owner (public key address). From this, and the public key of the Mirror instance, we can compute the address of the state machine script that locks the token for the owner.

Constructors

CredentialOwnerReference

Fields

coTokenName :: TokenName

coOwner :: Wallet

data MirrorError

Constructors

StateNotFound TokenName PaymentPubKeyHash

SetupError ContractError

MirrorEndpointError ContractError

CreateTokenTxError ContractError

StateMachineError SMContractError

mirror :: (HasEndpoint "revoke" CredentialOwnerReference s, HasEndpoint "issue" CredentialOwnerReference s) => Contract w s MirrorError ()