Plutus.Contracts.Prism.StateMachine

Description

State machine that manages credential tokens

Documentation

data IDState

Constructors

Active
The credential is active and can be used in transactions

Revoked
The credential has been revoked and can't be used anymore.

data IDAction

Constructors

PresentCredential
Present the credential in a transaction

RevokeCredential
Revoke the credential

data UserCredential

A Credential issued to a user (public key address)

Constructors

UserCredential

Fields

ucAddress :: PaymentPubKeyHash

Address of the credential holder

ucCredential :: Credential

The credential

ucToken :: Value

The Value containing a token of the credential (this needs to be included here because token is not available in on-chain code)

typedValidator :: UserCredential -> TypedValidator (StateMachine IDState IDAction)

machineClient :: TypedValidator (StateMachine IDState IDAction) -> UserCredential -> StateMachineClient IDState IDAction

mkMachineClient :: CredentialAuthority -> PaymentPubKeyHash -> TokenName -> StateMachineClient IDState IDAction