Plutus.Contracts.SealedBidAuction

Documentation

data AuctionParams

Definition of an auction

Constructors

AuctionParams

Fields

apOwner :: Address

Current owner of the asset. This is where the proceeds of the auction will be sent.

apAsset :: Value

The asset itself. This value is going to be locked by the auction script output.

apEndTime :: POSIXTime

When the time window for bidding ends.

apPayoutTime :: POSIXTime

When the time window for revealing your bid ends.

newtype BidArgs

Constructors

BidArgs

Fields

secretBid :: SecretArgument Integer

newtype RevealArgs

Constructors

RevealArgs

Fields

publicBid :: Integer

data AuctionError

Constructors

StateMachineContractError SMContractError
State machine operation failed

AuctionContractError ContractError
Endpoint, coin selection, etc. failed

type BidderSchema = (Endpoint "bid" BidArgs .\/ Endpoint "reveal" RevealArgs) .\/ Endpoint "payout" ()

type SellerSchema = Endpoint "payout" ()

startAuction :: Value -> POSIXTime -> POSIXTime -> Contract () SellerSchema AuctionError ()

bid :: AuctionParams -> Promise () BidderSchema AuctionError ()

reveal :: AuctionParams -> Promise () BidderSchema AuctionError ()

payout :: HasEndpoint "payout" () s => AuctionParams -> Promise () s AuctionError ()

packInteger :: Integer -> BuiltinByteString

Pack an integer into a byte string with a leading sign byte in little-endian order

sellerContract :: AuctionParams -> Contract () SellerSchema AuctionError ()

Top-level contract for seller

bidderContract :: AuctionParams -> Contract () BidderSchema AuctionError ()

Top-level contract for buyer