Plutus.Contracts.SimpleEscrow

Description

This simple escrow contract facilitiates and exchange of currencies.

Documentation

data EscrowParams

Constructors

EscrowParams

Fields

payee :: PaymentPubKeyHash

The entity that needs to be paid the expecting Value.

paying :: Value

Value to be paid out to the redeemer.

expecting :: Value

Value to be received by the payee.

deadline :: POSIXTime

Time after which the contract expires.

type EscrowSchema = (Endpoint "lock" EscrowParams .\/ Endpoint "refund" EscrowParams) .\/ Endpoint "redeem" EscrowParams

data Action

Constructors

Redeem

Refund

data RedeemFailReason

Constructors

DeadlinePassed

data EscrowError

Constructors

RedeemFailed RedeemFailReason

RefundFailed

EContractError ContractError

class AsEscrowError r where

Minimal complete definition

_EscrowError

Methods

_EscrowError :: Prism' r EscrowError

_RedeemFailed :: Prism' r RedeemFailReason

_RefundFailed :: Prism' r ()

_EContractError :: Prism' r ContractError

newtype RefundSuccess

Constructors

RefundSuccess TxId

newtype RedeemSuccess

Constructors

RedeemSuccess TxId

data Escrow

escrowAddress :: CardanoAddress

escrowInstance :: TypedValidator Escrow

validate :: EscrowParams -> Action -> ScriptContext -> Bool

lockEp :: Promise () EscrowSchema EscrowError ()

Lock the paying Value in the output of this script, with the requirement that the transaction validates before the deadline.

redeemEp :: Promise () EscrowSchema EscrowError RedeemSuccess

Attempts to redeem the Value locked into this script by paying in from the callers address to the payee.

refundEp :: Promise () EscrowSchema EscrowError RefundSuccess

Refunds the locked amount back to the payee.