Plutus.Contracts.Stablecoin

Documentation

newtype SC a

An amount of stablecoins

Constructors

SC

Fields

unSC :: a

newtype RC a

An amount of reservecoins

Constructors

RC

Fields

unRC :: a

newtype BC a

An amount of base currency coins (eg. Ada or some native currency)

Constructors

BC

Fields

unBC :: a

newtype PC a

An amount of peg currency (eg. USD)

Constructors

PC

Fields

unPC :: a

data BankState

The bank's state

Constructors

BankState

Fields

bsReserves :: BC Integer

Value of the bank's reserves in base currency

bsStablecoins :: SC Integer

Amount of stablecoins in circulation

bsReservecoins :: RC Integer

Amount of reservecoins currently in circulation

bsMintingPolicyScript :: MintingPolicyHash

Hash of the minting policy that forwards all checks to the state machine. (This has to be in this type, rather than in Stablecoin, to avoid a circular dependency on the script's hash)

data Stablecoin

Stablecoin parameters.

Constructors

Stablecoin

Fields

scOracle :: PaymentPubKey

Public key of the oracle that provides exchange rates

scFee :: Rational

Fee charged by bank for transactions. Calculated as a fraction of the total transaction volume in base currency.

scMinReserveRatio :: Rational

The minimum ratio of reserves to liabilities

scMaxReserveRatio :: Rational

The maximum ratio of reserves to liabilities

scReservecoinDefaultPrice :: BC Integer

The price of a single reservecoin if no reservecoins have been issued

scBaseCurrency :: AssetClass

The asset class of the base currency. Value of this currency will be locked by the stablecoin state machine instance

scStablecoinTokenName :: TokenName

TokenName of the stablecoin

scReservecoinTokenName :: TokenName

TokenName of the reservecoin

data Input

Input to the stablecoin state machine

Constructors

Input

Fields

inpSCAction :: SCAction

The action to be performed

inpConversionRate :: SignedMessage (Observation ConversionRate)

Exchange rate between base currency and peg currency

data SCAction

Action that can be performed on the stablecoin contract.

Constructors

MintStablecoin (SC Integer)
Create a number stablecoins, depositing the matching amount of base currency

MintReserveCoin (RC Integer)
Create a number of reservecoins, depositing the matching amount of base currency

type ConversionRate = Rational

Conversion rate from peg currency (eg. USD) to base currency (eg. Ada)

State machine client

typedValidator :: Stablecoin -> TypedValidator (StateMachine BankState Input)

machineClient :: TypedValidator (StateMachine BankState Input) -> Stablecoin -> StateMachineClient BankState Input

step :: forall i o. Stablecoin -> BankState -> Input -> Maybe (TxConstraints i o, BankState)

Contract using the state machine

contract :: Promise () StablecoinSchema StablecoinError ()

A Contract that initialises the state machine and then accepts Input transitions.

data StablecoinError

type StablecoinSchema = Endpoint "run step" Input .\/ Endpoint "initialise" Stablecoin

Etc.

stableCoins :: Stablecoin -> SC Integer -> Value

A Value with the given number of stablecoins

reserveCoins :: Stablecoin -> RC Integer -> Value

A Value with the given number of reservecoins

checkValidState :: Stablecoin -> BankState -> ConversionRate -> Either InvalidStateReason ()