Plutus.Contracts.TokenAccount

Description

Plutus implementation of an account that can be unlocked with a token. Whoever owns the token can spend the outputs locked by the contract. (A suitable token can be created with the Currency contract, or with newAccount in this module)

Documentation

newtype Account

Constructors

Account

Fields

accountOwner :: AssetClass

Contract functionality

pay :: AsTokenAccountError e => Account -> Value -> Contract w s e CardanoTx

Pay some money to the given token account

redeem

Arguments

:: AsTokenAccountError e

=> CardanoAddress
Where the token should go after the transaction

-> Account
The token account

-> Contract w s e CardanoTx

Empty the account by spending all outputs belonging to the Account.

newAccount

Arguments

:: forall w s e. AsTokenAccountError e

=> TokenName
Name of the token

-> CardanoAddress
Address of the token's initial owner

-> Contract w s e Account

Create a new token and return its Account information.

balance :: AsTokenAccountError e => Account -> Contract w s e Value

balance account returns the value of all unspent outputs that can be unlocked with accountToken account

address :: Account -> CardanoAddress

accountToken :: Account -> Value

payTx :: Value -> TxConstraints (RedeemerType TokenAccount) (DatumType TokenAccount)

A transaction that pays the given value to the account

redeemTx :: forall w s e. AsTokenAccountError e => Account -> CardanoAddress -> Contract w s e (TxConstraints () (), ScriptLookups TokenAccount)

Create a transaction that spends all outputs belonging to the Account.

Endpoints

type TokenAccountSchema = (Endpoint "redeem" (Account, CardanoAddress) .\/ Endpoint "pay" (Account, Value)) .\/ Endpoint "new-account" (TokenName, CardanoAddress)

type HasTokenAccountSchema s = (HasEndpoint "redeem" (Account, CardanoAddress) s, HasEndpoint "pay" (Account, Value) s, HasEndpoint "new-account" (TokenName, CardanoAddress) s)

tokenAccountContract :: forall w s e. (HasTokenAccountSchema s, AsTokenAccountError e) => Contract w s e ()

redeem, pay and newAccount with endpoints.

Etc.

data TokenAccount

data TokenAccountError

Constructors

TAContractError ContractError

TACurrencyError CurrencyError

class AsTokenAccountError r where

Minimal complete definition

_TokenAccountError

Methods

_TokenAccountError :: Prism' r TokenAccountError

_TAContractError :: Prism' r ContractError

_TACurrencyError :: Prism' r CurrencyError

validatorHash :: Account -> ValidatorHash

typedValidator :: Account -> TypedValidator TokenAccount