Plutus.Contracts.Uniswap.OffChain

Documentation

poolStateCoinFromUniswapCurrency

Arguments

:: CurrencySymbol
The currency identifying the Uniswap instance.

-> Coin PoolState

Gets the Coin used to identity liquidity pools.

liquidityCoin

Arguments

:: CurrencySymbol
The currency identifying the Uniswap instance.

-> Coin A
One coin in the liquidity pair.

-> Coin B
The other coin in the liquidity pair.

-> Coin Liquidity

Gets the liquidity token for a given liquidity pool.

data CreateParams

Parameters for the create-endpoint, which creates a new liquidity pool.

Constructors

CreateParams

Fields

cpCoinA :: Coin A

One Coin of the liquidity pair.

cpCoinB :: Coin B

The other Coin.

cpAmountA :: Amount A

Amount of liquidity for the first Coin.

cpAmountB :: Amount B

Amount of liquidity for the second Coin.

data SwapParams

Parameters for the swap-endpoint, which allows swaps between the two different coins in a liquidity pool. One of the provided amounts must be positive, the other must be zero.

Constructors

SwapParams

Fields

spCoinA :: Coin A

One Coin of the liquidity pair.

spCoinB :: Coin B

The other Coin.

spAmountA :: Amount A

The amount the first Coin that should be swapped.

spAmountB :: Amount B

The amount of the second Coin that should be swapped.

data CloseParams

Parameters for the close-endpoint, which closes a liquidity pool.

Constructors

CloseParams

Fields

clpCoinA :: Coin A

One Coin of the liquidity pair.

clpCoinB :: Coin B

The other Coin of the liquidity pair.

data RemoveParams

Parameters for the remove-endpoint, which removes some liquidity from a liquidity pool.

Constructors

RemoveParams

Fields

rpCoinA :: Coin A

One Coin of the liquidity pair.

rpCoinB :: Coin B

The other Coin of the liquidity pair.

rpDiff :: Amount Liquidity

The amount of liquidity tokens to burn in exchange for liquidity from the pool.

data AddParams

Parameters for the add-endpoint, which adds liquidity to a liquidity pool in exchange for liquidity tokens.

Constructors

AddParams

Fields

apCoinA :: Coin A

One Coin of the liquidity pair.

apCoinB :: Coin B

The other Coin of the liquidity pair.

apAmountA :: Amount A

The amount of coins of the first kind to add to the pool.

apAmountB :: Amount B

The amount of coins of the second kind to add to the pool.

type UniswapUserSchema = ((((((Endpoint "create" CreateParams .\/ Endpoint "swap" SwapParams) .\/ Endpoint "close" CloseParams) .\/ Endpoint "remove" RemoveParams) .\/ Endpoint "add" AddParams) .\/ Endpoint "pools" ()) .\/ Endpoint "funds" ()) .\/ Endpoint "stop" ()

Schema for the endpoints for users of Uniswap.

data UserContractState

Type of the Uniswap user contract state.

Constructors

Pools [((Coin A, Amount A), (Coin B, Amount B))]

Funds Value

Created

Swapped

Added

Removed

Closed

Stopped

type UniswapOwnerSchema = Endpoint "start" ()

start :: forall w s. Contract w s Text Uniswap

Creates a Uniswap "factory". This factory will keep track of the existing liquidity pools and enforce that there will be at most one liquidity pool for any pair of tokens at any given time.

create :: forall w s. Uniswap -> CreateParams -> Contract w s Text ()

Creates a liquidity pool for a pair of coins. The creator provides liquidity for both coins and gets liquidity tokens in return.

add :: forall w s. Uniswap -> AddParams -> Contract w s Text ()

Adds some liquidity to an existing liquidity pool in exchange for newly minted liquidity tokens.

remove :: forall w s. Uniswap -> RemoveParams -> Contract w s Text ()

Removes some liquidity from a liquidity pool in exchange for liquidity tokens.

close :: forall w s. Uniswap -> CloseParams -> Contract w s Text ()

Closes a liquidity pool by burning all remaining liquidity tokens in exchange for all liquidity remaining in the pool.

swap :: forall w s. Uniswap -> SwapParams -> Contract w s Text ()

Uses a liquidity pool two swap one sort of coins in the pool against the other.

pools :: forall w s. Uniswap -> Contract w s Text [((Coin A, Amount A), (Coin B, Amount B))]

Finds all liquidity pools and their liquidity belonging to the Uniswap instance. This merely inspects the blockchain and does not issue any transactions.

ownerEndpoint :: Contract (Last (Either Text Uniswap)) EmptySchema ContractError ()

userEndpoints :: Uniswap -> Promise (Last (Either Text UserContractState)) UniswapUserSchema Void ()

Provides the following endpoints for users of a Uniswap instance:

create

Creates a liquidity pool for a pair of coins. The creator provides liquidity for both coins and gets liquidity tokens in return.

swap

Uses a liquidity pool two swap one sort of coins in the pool against the other.

close

Closes a liquidity pool by burning all remaining liquidity tokens in exchange for all liquidity remaining in the pool.

remove

Removes some liquidity from a liquidity pool in exchange for liquidity tokens.

add

Adds some liquidity to an existing liquidity pool in exchange for newly minted liquidity tokens.

pools

Finds all liquidity pools and their liquidity belonging to the Uniswap instance. This merely inspects the blockchain and does not issue any transactions.

funds

Gets the caller's funds. This merely inspects the blockchain and does not issue any transactions.

stop

Stops the contract.

findSwapA :: Amount A -> Amount B -> Amount A -> Integer

findSwapB :: Amount A -> Amount B -> Amount B -> Integer

covIdx :: CoverageIndex

findUniswapFactoryAndPool :: forall w s. Uniswap -> Coin A -> Coin B -> Contract w s Text ((TxOutRef, DecoratedTxOut, [LiquidityPool]), (TxOutRef, DecoratedTxOut, LiquidityPool, Amount Liquidity))

uniswapInstance :: Uniswap -> TypedValidator Uniswapping

liquidityPolicy :: Uniswap -> MintingPolicy

uniswapScript :: Uniswap -> Validator

poolStateCoin :: Uniswap -> Coin PoolState

liquidityCurrency :: Uniswap -> CurrencySymbol

lpTicker :: LiquidityPool -> TokenName

Generate a unique token name for this particular pool; based on the tokens it exchanges. This should be such that looking for a pool exchanging any two tokens always yields a unique name.

calculateRemoval :: Amount A -> Amount B -> Amount Liquidity -> Amount Liquidity -> (Amount A, Amount B)

See Definition 3 of https://github.com/runtimeverification/verified-smart-contracts/blob/c40c98d6ae35148b76742aaaa29e6eaa405b2f93/uniswap/x-y-k.pdf.

funds :: forall w s. Contract w s Text Value

Gets the caller's funds.