Plutus.Contracts.Uniswap.Types

Documentation

data U

Uniswap coin token

Constructors

U

data A

A-side coin token

Constructors

A

data B

B-side coin token

Constructors

B

data PoolState

Pool-state coin token

Constructors

PoolState

data Liquidity

Liquidity-state coin token

Constructors

Liquidity

newtype Coin a

A single AssetClass. Because we use three coins, we use a phantom type to track which one is which.

Constructors

Coin

Fields

unCoin :: AssetClass

newtype Amount a

Likewise for Integer; the corresponding amount we have of the particular Coin.

Constructors

Amount

Fields

unAmount :: Integer

valueOf :: Coin a -> Amount a -> Value

unitValue :: Coin a -> Value

isUnity :: Value -> Coin a -> Bool

amountOf :: Value -> Coin a -> Amount a

mkCoin :: CurrencySymbol -> TokenName -> Coin a

newtype Uniswap

Constructors

Uniswap

Fields

usCoin :: Coin U

data LiquidityPool

Constructors

LiquidityPool

Fields

lpCoinA :: Coin A

lpCoinB :: Coin B

data UniswapAction

Constructors

Create LiquidityPool

Close

Swap

Remove

Add

data UniswapDatum

Constructors

Factory [LiquidityPool]

Pool LiquidityPool (Amount Liquidity)