Plutus.Contracts.Vesting

Documentation

data VestingParams

A vesting scheme consisting of two tranches. Each tranche defines a date (POSIX time) after which an additional amount can be spent.

Constructors

VestingParams

Fields

vestingTranche1 :: VestingTranche

vestingTranche2 :: VestingTranche

vestingOwner :: PaymentPubKeyHash

type VestingSchema = Endpoint "vest funds" () .\/ Endpoint "retrieve funds" Value

A simple vesting scheme. Money is locked by a contract and may only be retrieved after some time has passed.

This is our first example of a contract that covers multiple transactions, with a contract state that changes over time.

In our vesting scheme the money will be released in two _tranches_ (parts): A smaller part will be available after an initial number of time has passed, and the entire amount will be released at the end. The owner of the vesting scheme does not have to take out all the money at once: They can take out any amount up to the total that has been released so far. The remaining funds stay locked and can be retrieved later.

Let's start with the data types.

data VestingTranche

Tranche of a vesting scheme.

Constructors

VestingTranche

Fields

vestingTrancheDate :: POSIXTime

vestingTrancheAmount :: Value

data VestingError

Constructors

VContractError ContractError

InsufficientFundsError Value Value Value

class AsVestingError r where

Minimal complete definition

_VestingError

Methods

_VestingError :: Prism' r VestingError

_VContractError :: Prism' r ContractError

_InsufficientFundsError :: Prism' r (Value, Value, Value)

totalAmount :: VestingParams -> Value

The total amount vested

vestingContract :: VestingParams -> Contract () VestingSchema VestingError ()

validate :: VestingParams -> () -> () -> ScriptContext -> Bool

vestingScript :: VestingParams -> Validator